+++ package.json
@@ -31,6 +31,7 @@
     "gulp-autoprefixer": "^8.0.0",
     "gulp-clean": "^0.4.0",
     "gulp-clean-css": "^4.3.0",
+    "gulp-concat": "^2.6.1",
     "gulp-rename": "^2.0.0",
     "gulp-sass": "^5.1.0",
     "gulp-sourcemaps": "^3.0.0",
@@ -42,7 +43,6 @@
   },
   "dependencies": {
     "@fancyapps/ui": "^5.0.16",
-    "gulp-concat": "^2.6.1",
     "swiper": "^9.2.0"
   }
 }
