+++ src/index.html
@@ -706,8 +706,6 @@
                                     <div class="card-master__experience">опыт работы 5 лет</div>
                                 </div>
                             </div>
-                            </в>
-
                         </div>
                     </div>
             </section> <!-- /masters -->
