+++ src/js/tabs/index.js
@@ -1,0 +1,8 @@
+import TabsManager from './TabsManager';
+
+const tabs = () => {
+    const tabsElem = document.querySelector('#tabs-section');
+    new TabsManager(tabsElem);
+}
+
+export default tabs;
