+++ src/sass/blocks/footer.scss
@@ -1,8 +1,6 @@
 .footer {
     margin-top: auto;
-}
 
-.footer {
     &__inner {
         display: flex;
         justify-content: space-between;
