+++ src/sass/blocks/header.scss
@@ -1,11 +1,3 @@
-.header {
-    width: 100%;
-    height: auto;
-
-    background: url('../images/dest/header/header-bg.jpg') center no-repeat;
-    background-size: cover;
-}
-
 .logo {
     font-weight: 700;
     font-size: 24px;
@@ -20,9 +12,12 @@
 
 .header {
     position: relative;
-}
 
-.header {
+    width: 100%;
+    height: auto;
+
+    background: url('../images/dest/header/header-bg.jpg') center no-repeat;
+    background-size: cover;
 
     &__top {
         position: relative;
